+++ src/index.js
@@ -22,7 +22,7 @@ import Intro from './components/intro.jsx';
 import About from './components/about.jsx';
 import Portfolio from './components/portfolio.jsx';
 import Contact from './components/contact.jsx';
-import BackToTop from './components/back-top.jsx';
+import BackToTop from './components/footer.jsx';
 import Preloader from './components/preloader';
